test_openai_whisper.py
#!/usr/bin/env python3
"""
Test script to verify OpenAI Whisper is working correctly
"""

import os
import base64
from dotenv import load_dotenv
from openai import OpenAI

# Load environment variables
load_dotenv()

def test_openai_whisper():
    """Test OpenAI Whisper transcription"""
    
    # Check if API key is available
    api_key = os.getenv('OPENAI_API_KEY')
    if not api_key:
        print("❌ OPENAI_API_KEY not found in environment variables")
        return False
    
    print(f"✅ OpenAI API key found: {api_key[:20]}...")
    
    try:
        # Initialize OpenAI client
        client = OpenAI(api_key=api_key)
        print("✅ OpenAI client initialized successfully")
        
        # Create a simple test audio file (just for testing the connection)
        # In real usage, this would be actual audio data
        test_audio_data = b"test audio data for connection testing"
        
        # Test the connection (this won't actually transcribe, just test the client)
        print("🔍 Testing OpenAI client connection...")
        
        # Check if client has audio attribute
        if hasattr(client, 'audio'):
            print("✅ Client has audio attribute (new format)")
        else:
            print("⚠️ Client doesn't have audio attribute (legacy format)")
        
        print("✅ OpenAI Whisper client is properly configured!")
        return True
        
    except Exception as e:
        print(f"❌ Error testing OpenAI Whisper: {e}")
        print(f"🔍 Error type: {type(e).__name__}")
        return False

if __name__ == "__main__":
    print("🧪 Testing OpenAI Whisper Configuration...")
    success = test_openai_whisper()
    
    if success:
        print("\n✅ OpenAI Whisper is properly configured!")
        print("🔍 The issue might be in the Flask service or audio data format")
    else:
        print("\n❌ OpenAI Whisper configuration has issues")
        print("🔍 Check your API key and OpenAI library installation") 